refactor(ai-impact): move pipeline phases to main nav sidebar
Replace the in-page PhaseSidebar with top-level nav items in the app
sidebar. Each phase/workflow is now a separate view with its own route,
matching the pattern used by other modules (team-tracker, etc.).

- Split monolith AIImpactView into RFEReviewView, FeatureReviewView,
  and AutofixView
- Add disabled nav items for coming-soon phases with separator before
  workflows section
- Support `disabled` and `separatorBefore` properties on navItems in
  AppSidebar
- Cross-linking between views uses moduleNav.navigateTo with select
  param
